requirements.txt
Flask==2.2.5
Werkzeug<3.0
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
pyjwt==2.7.0
bcrypt==4.0.1
pytest==7.4.0